feat(pre-aggregation): Expose charge properties in flat filters (#4034)
## Context

This PR is part of the Pre-aggregation epic.

The main goal of this feature is to setup a complete aggregation
pipeline that will be leveraged to compute the customer usage without
re-querying the full set of events.

## Description

This pull request is exposing the `pricing_group_keys` and `properties`
in the flat_filters view, to allow the events_processor to use them when
enriching the events.