fix(activity-log): Do not include fees when there are too many of them (#4102)
## Context

There is an issue when we're calling `ActivityLog` with a huge payload
e.g. an invoice with a lot of fees.
In that case we're hitting the Kafka payload limit (1MB by default) and
getting `WaterDrop::Errors::MessageInvalidError`

## Description

In order to fix this issue we skip fees to be serialized for invoices
that have more than 25 fees.
So that `ActivityLog` would still log the invoice except the fees in
this case.