fix(charge-model-factory): Boolean param to only calculate the projected values when being called by the projection service (#4118)
## Context

Related to the RAM increase after the projected usage feature was
merged. This one complements the previous that added a boolean flag to
only calculate the projected values from the right parts of the
application.

## Description

In the `ChargeModelFactory`, added a boolean attribute that is only
`true` when coming from the `ProjectionService`.