misc: Add BillingPerdiodBoundaries (#4121)
## Context

In many places of the application, we are defining a `boundaries`
object, sometimes as an instance of a `Hash`, sometimes as an
`OpenStruct`. The fields are not always the same and since the
boundaries are injected in a lot of services, this causes some
uncertainty when accessing the properties.

## Description

In an attempt to improve the readability and the reliability of this key
part of the application, the goal of this PR is to introduce a new
`BillingPeriodBoundaries` class and to use it everywhere in replacement
of `Hash` and `OpenStruct`.

It should also improve the memory consumption as OpenStructs are known
to cause memory issues in Ruby when YJIT is enabled