misc(event): remove unique validation (#4133)
## Context

This PR intend to improve the performance when creating an event using
the `POST /api/v1/events`.

The event model has a uniqueness validation on the transaction_id field.
Rails is performing a database query to validate this uniquess and it
appears to be slow...

## Description

Since we already have a unique index in place in the database, this PR
removes the validation and adds the logic to catch the Postgres error.
Some load testing have been conducted with this approach and it cuts the
response time by a large factor, increasing the number of events/sec
that could be received by the API