feat(pre-aggregation): Expose charge pay_in_advance in flat filters (#4130)
## Context

This PR is part of the Pre-aggregation epic.

The main goal of this feature is to setup a complete aggregation
pipeline that will be leveraged to compute the customer usage without
re-querying the full set of events.

## Description

This pull request is exposing the `pay_in_advance` field in the
flat_filters view, to allow the events_processor use them when enriching
the events.