fix(clickhouse): Add more retry for lost connection (#4167)
## Context

This PR follows https://github.com/getlago/lago-api/pull/3733

## Description

In some situation, a "NoMethodError: undefined method 'select_all' for
nil" is raised when running
`::Clickhouse::EventsEnriched.connection.select_all`. This is an issue
with the Active record connector for Clickhouse.

This PR just handle an automatic retry when the error is raised