fix: advance charge fee when ingested right before billing run (#4082)
## Context

When pay in advance event that should be regrouped in the invoice is
ingested right before the billing run (e.g. 00:01:00); it got included
in the invoice for previous period. However, it should be included in
the invoice issued one month later.

## Description

This PR fixes described case, but also takes care that termination and
upgrade case stay untouched.