fix(timezones): Fix timezone names (#4177)
## Context

Following the update of Ruby to version 3.4.5, some issues where spotted
in the timezone especially when exporting the Timezone enum for the
GraphQL API.

It appears that some zone were renamed in ruby but that the support for
rails has not yet been changed accordingly.

## Description

This PR:
- Fixes the Timezone mapping by adding a new `Timezones` module that
will be used for both validations and GraphQL export.
- Add the correct handling for the impacted timezones:
  - Asia/Rangoon => Asia/Yangon
  - Europe/Kiev => Europe/Kyiv
  - America/Godthab => America/Nuuk
- Migrates the customers, organizations, invoices and billing entities
impacted by these changes