fix(events): Fix resolver with Clickhouse (#4182)
## Description
This PR is a fix for the following rails exception:

```
ActiveRecord::ActiveRecordError
Response code: 500: (ActiveRecord::ActiveRecordError)
["Code: 215. DB::Exception: Column `transaction_id` is not under aggregate function and not in GROUP BY. Have columns: ['count()']: While processing transaction_id, timestamp, external_subscription_id, code. (NOT_AN_AGGREGATE)"]
```

It's a regression introduced by a previous pull-request
(https://github.com/getlago/lago-api/pull/4163) that was made to
de-duplicates Clickhouse events in the GraphQL events resolver. The
issue is happening when querying an empty event page.