feat(entitlements): Introduce SubscriptionEntitlementQuery and remove view (#4176)
### New `SubscriptionEntitlementQuery`

This PR is a significant refactor of the SubscriptionEntitlement query.
The data model did not change, it's stored the same way but I have
redone the way we query the data subscriptions.
I removed the database view and replaced it with 2 queries, which
hydrates ActiveModel models (not ActiveRecord models). These models are
then serialized like any other model.

**100x improvement**: fixed all the bugs at once and made the thing 10x
more maintainable.

### New `SubscriptionEntitlementCreateOrUpdateService` service

The frontend will allow for each subscription feature to be updated
independently. The new services is tested through the GQP mutations but
it doesn't have proper tests yet. This is coming in a separate PR where
I will refactor services and more importantly: Add the abilitiy to
remove privilege from plan in subscription.

> [!IMPORTANT]  
> Remember the feature is not released yet.

## Upcoming next

- Merge API endpoints to "remove/restore" feature in subscription to
make it automatic
- Remove refactor services and add tests
- Add
`entitlement_subscription_feature_removals.entitlement_privilege_id` to
be able to remove privilege from subscription