feat(entitlement): Remove privilege (#4192)
As we allowed privilege to be removed in subscription in #4176 and #4190
, this PR ensure we correctly write the data from GQL and API.

The Services have been refactored so editing multiple features will loop
and call the "inner" service. The 2 "outer" services will trigger
webhooks, activitly logs, reformat errors...

Last part (in #4191) will delete the /remove and /restore endpoint and
add new better endpoints to remove privileges.

This was QA'ed with the frontend.