fix(entitlements): Convert boolean string for frontend (#4210)
With ActiveRecord if you do `EntitlementValue.update value: true`, it
will convert the value to `"t"` because the db column is string.
This is intentional and the serializer will use
`ActiveModel::Type::Boolean.new.cast("t")` to convert it to `true.

In GQL, we rely on `"true"` and `"false"` string to display the select
box. So we need to convert on read.
On write it's fine because the value will always be casted when returned
via the API.