misc: Add kosovo country code (#4213)
## Context

Lago relies on the https://github.com/countries/countries gem to get the
list of available country. It is using the list of ISO 3166 country
codes.

Some customers recently ask for the inclusion of the Kosovo in the list
of countries. The issue is that for now, no ISO 3166 country code is
assigne to the Kosovo.

An issue is opened on the countries repository about it:
https://github.com/countries/countries/issues/793
I appears that it will not be fixed until the Kosovo officially received
an ISO 3166 country code.

## Description

To address the request from customers, this PR is using the proposed
workaround from the source repository issue.
The PR also updates the `countries` gem