fix(event): Kafka payload for batch events (#4220)
## Context

Kafka payload for raw events is built in two places, one in the
`Events::CreateService` and one in `Events::CreateBatchService` this
issue is that the one from the batch service is lacking the `metadata`
fields.

This field is a hash used by the events processor to find if the event
is already post processed in the API. If not and if the event matched an
in advance charge, a message is produced in the
`events_charged_in_advance` topic.

As the metadata was missing the event processor was considering that the
event was not post-processed by the API even if it was, leading to a
double processing and a duplicated fee...

## Description

This PR makes sure that both `Events::CreateService` and
`Events::CreateBatchService` uses the same Kafka producer service to
remove the duplicated payload handling.

NOTE: The duplication should have been avoided by the unique index
defined at Posgres level but another bug have been identified at this
level and will be fixed in a second PR.