feat(pre-aggregation): Pass charge_id and charge_filter_id to the event stores (#4231)
## Context

This PR is part of the Pre-aggregation epic.

The main goal of this feature is to setup a complete aggregation
pipeline that will be leveraged to compute the customer usage without
re-querying the full set of events.

## Description

This pull request pass the `charge_id` and `charge_filter_id` to the
event stores. They will be used later to filter events and
pre-aggregated events