fix(alerting): When BillableMetric is deleted, delete associated alerts (#4125)
Alerts can be specific to a billable metric. When a billable metric is
deleted, all associated alerts should also be deleted since they cannot
be triggered anymore.

I changed the relationship to ensure that serializing a triggered alert
correctly retrieve the alert and the billable metric even if it's
deleted. That's the point of soft deletion, right?
It's possible that the `alert.triggered` webhook is sent after the alert
was deleted.

Currently, if you delete an BM and you had alerts associated with it.
the `/subscriptions/:external_id/alerts` will throw a 500 because we
cannot serialize it.

---------

Co-authored-by: Ancor Cruz <hello@ancorcruz.com>