misc(clickhouse): Add clickhouse structure for cloud (#4202)
## Context
Today Cliickhouse cloud data structure structure is not stored anywhere
outside Clickhouse cloud itself this makes it a bit difficult to setup a
new instance when required.

## Description

This PR adds:
- All SQL scripts to create the database table and materialized view on
a Clickhouse cloud instance.
- Add the missing pre-aggregation logic on `precise_total_amount_cents` 

**NOTE**: the clickpipe configuration used to configure the Redpanda
consumers and the mapping of the fields from the topics to the table
fields is not exposed here.