test(invoice): Introduce invoice rendering tests (#4259)
## Context

I am currently working on a task that modifies the behavior of the invoice rendering for `credit` invoices. I noticed there were no spec for the rendering of those invoices. So I decided to add some tests and ended up extracting the logic into this PR.

## Description

In order to test the invoice rendering, I relied on the same code used by the `Utils::PdfGenerator`: `Slim::Template.new(template).render(invoice)`. The test is also using `build_stubbed` factory with constant data to have fast and consistent test setup. 

In order to make assertions on the view, I used the [`rspec-snapshots`](https://github.com/levinmr/rspec-snapshot) which serialize the HTML content into a snapshot on the first run and compares with the snapshot on the next runs. This allows us to easily diff the HTML changes and ensure there's no breaking change in the rendering.

One issue I faced though was that:

1. Slim rendering was not properly pretty printing tables:

   ```html
          <div class="mb-24">
            <h2 class="title-2 mb-4">
              $10.50
            </h2>
            <div class="body-1">
              Due Sep 04, 2025
            </div>
          </div>
          <div class="invoice-resume mb-24 overflow-auto">
            <table class="invoice-resume-table" width="100%"><tr><td class="body-2">Item</td><td class="body-2">Units</td><td class="body-2">Unit price</td><td class="body-2">Amount</td></tr><tr><td class="body-1">Prepaid credits - Premium Wallet</td><td class="body-2">10.5</td><td class="body-2">1.0</td><td class="body-2">$10.50</td></tr></table><table class="total-table" width="100%"><tr><td class="body-2"></td><td class="body-1">Total</td><td class="body-1">$10.50</td></tr></table>
          </div>
          
          <p class="body-3 mb-24">
            
          </p>
    ```

2. Nokogiri was not pretty printing with indentations:

    ```html
    <table class="total-table" width="100%"><tr>
    <td class="body-2"></td>
    <td class="body-1">Total</td>
    <td class="body-1">$10.50</td>
    </tr></table>
          </div>
          
          <p class="body-3 mb-24">
            
          </p>
          <div class="powered-by">
    <span class="body-2">Powered by  </span><img alt="Lago Logo" src="lago-logo-invoice.png">
    </div>
    ```
    
So I had to combine Slim with [`htmlbeautifier`](https://github.com/threedaymonk/htmlbeautifier/) gem to have a proper pretty print.