feat(pre-aggregation): Implement prorated_sum (#4258)
## Context

This PR is part of the Pre-aggregation epic, it follows
https://github.com/getlago/lago-api/pull/4236.

The main goal of this feature is to setup a complete aggregation
pipeline that will be leveraged to compute the customer usage without
re-querying the full set of events.

## Description

This PR implements:
- The base "events_sql" method, relying on the Clickhouse
events_enriched_expanded, used to build queries that cannot use the
`events_aggregated` table.
- The `prorated_sum`
- The `grouped_prorated_sum`
- The `sum_date_breakdown`