fix: refactor clickhouse query for prorated unique count (#4272)
## Context

There was a mismatch in some cases between postgres and clickhouse
results for prorated unique count.

## Description

To fix a mismatch, it was decided to take the postgres query as a source
of truth and rewrite clickhouse query.
Currently, we calculate period duration by converting timestamps to date
instead of handling seconds and dividing with number of seconds in the
day.
Also, some unnecessary rounding with ceil were removed.