feat(wallet-transaction): Add a `name` to wallet transactions (#4256)
## Roadmap Task

👉  https://getlago.canny.io/feature-requests/p/define-the-name-of-a-wallet-transaction-at-top-up

## Context

When creating a wallet transaction, the name displayed on the invoice defaults to the wallet name. However, for organizations managing multiple types of top-ups, it would be beneficial to display specific naming on invoices that corresponds to the particular transaction being made.

## Description 

This change adds a `name` attribute to the wallet transaction models, params and serializers.