fix(delete-tax): Preserve invoice information when tax is deleted (#4300)
## Problem
When we delete a `Tax` record, it deletes all the `AppliedTaxes` that
were created because of that `Tax`.
We do this for all `AppliedTaxes` records, regardless the owner of that
record. Finalized Invoice and Fee can't have that info deleted otherwise
they lose how they were calculated.

## Fix
`AppliedTax` and `Tax` relation has to be loosened, so `AppliedTax`
allow nil `Tax` reference.
Whenever we have a Tax deletion, we have to set
`Invoice::AppliedTax#tax` and `Fee::AppliedTax#tax` to `nil`.

We can only delete the `Invoice::AppliedTax#tax` and
`Fee::AppliedTax#tax` for draft invoices.
The other types of AppliedTaxes can be deleted because they are used on
execution time (i.e when the invoice is created/finalized).