fix(graphql): Fix date filters in API Logs and Activity Logs resolvers (#4275)
## Context

Offset pagination is fragile: if new record arrive between fetching page 1 and page 2, rows shift and clients either miss or re‑read items. This is what happened with the Activity Logs and API Logs feature for which we create new records more often that other features.

In order to mitigate this issue, the frontend is now always sending the `to_date` filter with the following rules:

1. If the filter is not set, we default to the timestamp at which we loaded the data for the first time
2. If the filter is set, we make sure that the value cannot be later than the timestamp at which we loaded the data for the first time

This ensures that no matter the requested page, we will never receive newer records.

The issue is that the backend GraphQL resolver did not accept `to_date` as a `ISO8601DateTime` but as a `ISO8601Date`. So the backend will convert the datetime into a date and back to a date time when querying clickhouse:

```ruby
lago-api(dev)> Clickhouse::ApiLog.where(logged_at: ..Date.today).to_sql
=> "SELECT api_logs.* FROM api_logs WHERE api_logs.logged_at <= '2025-09-11 00:00:00.000'"
```

## Description

This fixes it by adding a new `from_datetime` and `to_datetime`.

Note that it is note possible to:

1. Update `from_date` and `to_date` to `ISO8601DateTime` as it breaks the frontend
2. Using a union type as they do not support scalars.

It also adds tests for the resolvers filtering.