feat(wallets): follow Min and Max for transaction created by `target` rules (#4285)
When a target rule is processed we apply the minimum transaction amount
on the wallet, but not the maximum.

The `gap` is the different between the current balance and the target
balance.

**Examples**
 
```
- current balance: 12
- target balance: 30
- gap: 18
- wallet minimum: 25
```
In this case, we'll create a transaction of 25 credits putting the
**balance to 37, instead of 30**.

```
- current balance: 12
- target balance: 30
- gap: 18
- wallet maximum: 10
```
In this case, we'll ignore the maximum and create a transaction of 18
credits to put the balance to target. Reaching target is more important
than maximum.

---------

Co-authored-by: Oleksandr Chebotarov <oleksandr@getlago.com>