misc: Graduated prorated refactoring (#4199)
## Context

We decided to change a bit graduated prorated logic.

## Description

Before; added and removed unit was treated as 1 unit with value prorated
by the number of days between addition and removal.

After this change, removed unit is not treated as 1 unit but instead as
zero unit and it does not result in changing the tier. However, its
prorated value is still used inside the tier and it affects the final
result.