fix(billable_metric_filter): Fix batch remove (#4328)
## Context

This PR fixes the `Discard::RecordNotDiscarded - A discarded record
cannot be discarded (Discard::RecordNotDiscarded)` that is raised in
`BillableMetricFilters::CreateOrUpdateBatchService` when try to remove
all filters on a billable metric.

The issue was raised because the same billable_metric_filter was used
multiple time on the same charge filter.

## Description

This PR makes sure that the filter value that we are trying to soft
delete is not already soft deleted, to avoid raising the exception