feat(wallet-transaction): Add `transaction_name` to recurring transaction rules (#4282)
## Roadmap Task

👉  https://getlago.canny.io/feature-requests/p/define-the-name-of-a-wallet-transaction-at-top-up

## Context

When creating a wallet transaction, we can set a `name` that will be displayed on invoice. But this is not yet the case for wallet transactions created by recurring rules.

## Description

This adds a `transaction_name` to the wallet transaction recurring rule that are then used to populate the name of the triggered wallet transactions.