fix: Ability to update a subscription without passing bm_id (#4367)
## Description

Updating a subscription payload is not aligned with creating one.

Creating one only requires an `id` for charge override. However, the
update one requires both the `id` **AND** the `billable_metric_id`.

The goal of this PR is to be able to update an existing subscription by
passing a charge without its `billable_metric_id`.