fix(progressive_billing): Avoid parallel processing (#4361)
## Context

Sometimes `UsageMonitoring::ProcessSubscriptionActivityService` and
`LifetimeUsages::RecalculateAndCheckJob` are performed at the same time.

This leads to a lot of `Idempotency::IdempotencyError` as a logic to
avoid duplicating invoices is in place to prevent the same usage
threshold from being billed twice.
In some rare situtation (when we are out of luck) it could lead to a
double invoicing of the same thresold, with the amount of the first
invoice immediately re-credited on the second one. This situation is
related to a race condition appearing when the
`LifetimeUsages::CalculateService` start to process just before the
commit of the Idempotent transaction.

It appears that the `Idempotent.transaction` is not enough to mitigate
this situation as it does not cover the full computation leading to a
Progressive Billing invoice.

## Description

This PR re-use the existing `LifetimeUsages::RecalculateAndCheckJob` in
the `UsageMonitoring::ProcessSubscriptionActivityService` in order to
take advantage of the `active_job_uniqueness` making sure that we cannot
process multiple check for the same lifetime usage record in parallel.

The job is enqueued using `perform_now` because it takes the
current_usage as an input parameter and we don't want to add new load
with serialization/deserialization of huge JSON objects to this
pipeline.