fix(wallet): Fix behavior when `invoice_requires_successful_payment` is nil (#4372)
## Context

Because `ActiveModel::Type::Boolean.new.cast(nil)` returns `nil` and the `invoice_requires_successful_payment` has a `NOT NULL` db constraint, we fail to create a wallet or wallet transaction when `invoice_requires_successful_payment` is set to `nil`.

## Description

This fixes this behavior by defaulting to `false` when creating a wallet and defaulting to the wallet's value when creating a transaction. Validations were also added on the model to ensure saving such value will fail.