Feat(fixed_charges): add events aggregators (#4320)
## Context

For every subscription, when a subscription starts or when units for a
fixed charge are updated, we'll be issuing the FixedChargeEvent. Later,
when calculating the fixedChargesFees, we need to aggregate these events
(prorated and not) in order to correctly calculate the fees. This PR
introduces the aggregation services, that return units amount, onto
which later we'll be applying fixed_charges charge_models to get the
resulting fee amount

## Description

- Introduced FixedChargeEvent aggregation services