test(factory): Add `with_static_values` trait to some factories (#4371)
## Context

In order to test invoice rendering, we are using static values for customer, organization and billing entities. These values are not really re-usable as is. If we want to add more invoice rendering tests, we'd have to copy-paste those values.

## Description

This adds a `with_static_values` trait to the customer, organization and billing entity factories which sets fixed values for attributes used in the invoice rendering. We can then easily re-use those factories with:

```ruby
let(:customer) { create(:customer, :with_static_values) }
```