test: Improve `match_html_snapshot` matcher (#4376)
## Context

We rely on the `match_html_snapshot` to test the invoice rendering. One issue with it is that it includes a huge chunk of CSS and unnecessary `<head>` which makes the snapshot and diff bigger than necessary.

## Description

The matcher will now strip the `<head>` and `<style>` tags by default and ensure each HTML tag is on a new line to ensure consistent formatting.