fix(credit_notes): Handle rounding issues with taxes (#4373)
## Context

This PR is a fix for some rounding issues that have been spotted on
credit notes:

## Here is an example scenario

- Create a subscription starting on 2025/09/19 for a monthly plan with a
subscription billed in advance for 200$ with 25% VAT. The plan is
renewed on an anniversary basis.
  - An invoice for 250$ is emitted 

- Register a full payment on the invoice (for 250$)

- On the same day, terminate the subscription.
  - A credit note will be emitted with these values:
    - sub total before taxes: 193.33
    - taxes amount: 48.33 
    - total amount: 241.67 <== If we sum, this should be  241.66

## Description

The reason behind this issue is that different rounding are applied when
creating a credit note (on fees, applied coupons, taxes, credit and
refund amount).
The difference will always be +/-1$ around the expected amount.

This PR add a new logic to adjust the total, credit, refund and taxes
amount after computation to make sure it always make sense from a
customer perspective. The rounding of taxes and total will be adjusted
with a "rounding adjustment".
The logic to take this adjustment into account in case of successive
credit notes already exists.